fix: serialize inventory count as a number

+++ biyesheji-common/src/main/java/com/biyesheji/vo/MerchantInventorySummaryVO.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class MerchantInventorySummaryVO {
-    private long lowStockCount;
+    private int lowStockCount;
     private int threshold;
 }
