+++ product-service/src/main/java/com/biyesheji/product/mapper/InventoryMapper.java
@@ -38,5 +38,5 @@ public interface InventoryMapper {
             JOIN t_stock stock ON stock.sku_id = sku.id
             WHERE p.deleted = 0 AND p.status = 1 AND sku.status = 1 AND stock.available <= #{threshold}
             """)
-    long countLowStock(@Param("threshold") int threshold);
+    int countLowStock(@Param("threshold") int threshold);
 }
